+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -9,7 +9,7 @@ public class Calculator {
         public int subtract(int a, int b) {
             return a - b;
     }
-    public int square(int a) {
+    public double square(double a) {
         return a * a;
     }
